STABLE-6895: Part 2/3, implement depositForBurnWithHook relay-side  (#13)
## Summary

This implements `depositForBurnWithHook` on the relay side. 

This includes:
- Updating `TokenMessengerV2` with a new `depositForBurnWithHook`
function
- Validating inputs 
- Using `BurnMessageV2` to format the burn message correctly (note: the
Iris-inserted fields are omitted).
- Calling message transmitter with the formatted message.

It does _not_ cover:
- MessageTransmitt-ing 
- Handling receive messages or executing hooks 

^ will be taken care of in part 3. 

## Testing 

This extends the `TokenMessengerV2` tests, using many of the same test
cases from `TokenMessenger.t.sol`. Some callouts and questions below
in-line.

## Discussion

Open to any and all ideas for further code re-use between v1 and v2. The
principle has been to leave v1 as-is, so re-use as focused on
implementations and tests, vs strictly sharing code.