STABLE-6895 (Part 4) / STABLE-6926: Add deterministic nonce handling in MessageTransmitterV2 (#19)
### Summary

This adds deterministic nonce handling in MessageTransmitterV2. It
completes the `receiveMessage` implementation and adds a bunch of tests
derived from the v1 testcases.

A few callouts below in-line. The vast majority of the diff is in tests
and license headers for new files--the actual contract changes are
fairly compact.

### Testing

See `v2/MessageTransmitterV2.t.sol` for test additions. Please
scrutinize the test cases and compare them to the v1 tests in
`MessageTransmitter.t.sol` and see if anything is missing. I opted to
focus on fuzzing additional attributes and using mocks to focus on the
contract under test.