STABLE-7589: Add Denylist to TokenMessenger (#31)
### Summary

Adds a Denylist to TokenMessenger. Updated the ABI to align with the
Paymaster PR:
https://github.com/circlefin/smart-wallet-contracts-private/pull/155,
though the implementations are slightly difference.

The requirements are to check the denylist on `depositForBurn` and
`depositForBurnWithHook` for both the `msg.sender` and `tx.origin`
addresses.

This is pretty straightforward, but some callouts are down below. 

### Testing

Added dedicated tests using a mock Denylistable, and unit tests for
TokenMessengerV2 and BaseTokenMessenger.