STABLE-6894: Part 2, complete proxy implementation (#30)
### Summary

This adds an `Initializable` type and updates `TokenMessengerV2` and
`MessageTransmitterV2` to use it. It also adds initializers to both
contracts.

Some implementation notes:
- The `Initializable` type is forked from OZ, and then modified for
Solidity 0.7.6 (no custom errors) and stripped down of features that
we're not using.
- Please scrutinize the fork changes to `Initializable`. I'm not wed to
using this type for our purposes.
- Had to create internal setters for Pausable, Rescuable, and Attestable
to allow the initializers to set those values.

Some further callouts down below

### Testing

Added tests for Initializable + the initializer function additions to
the TokenMessenger and MessageTransmitter, and also simulated an upgrade
to mock v3 versions.