[STABLE-8092] [Cont.]: Update implementation deployment script to use CREATE2 more effectively (#60)
### Summary 

Follow-up to:
https://github.com/circlefin/evm-cctp-contracts-private/commit/8cf36730033099b02d2ab06a27faaf4e8c130a74,
this correct an oversight where the TokenMinterV2 via Create2 was being
passed the tokenController as a constructor arg. Since this affects the
bytecode, it affects the address.

### Testing

The tests actually already assert that the tokenController is rotated
correctly; it's done atomically in both ways, only this way ensures a
consistent address.

Edit: added a test to ensure that the script is deploying via CREATE2 in
an expected way.